+++ packages/generator-teams/CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Updated build packages
 * Moved custom `skip-install` implementation to native Yeoman version
 * Updated to Microsoft Teams JS SDK 1.10
+* Node 10.x no longer supported
 
 ### Deleted